+++ src/CMakeLists.txt
@@ -44,12 +44,6 @@ if (PCMS_ENABLE_OMEGA_H)
           pcms/uniform_grid.h
           pcms/point_search.h)
 endif ()
-if (PCMS_ENABLE_SERVER)
-  if (NOT PCMS_ENABLE_OMEGA_H)
-    message(ERROR "PCMS_ENABLE_OMEGA_H is required for server implementation")
-  endif ()
-  list(APPEND PCMS_HEADERS pcms/server.h)
-endif ()
 
 find_package(Kokkos REQUIRED)
 find_package(perfstubs REQUIRED)
