+++ test/test_proxy_coupling.cpp
@@ -2,7 +2,6 @@
 #include <iostream>
 #include <pcms.h>
 #include <pcms/types.h>
-#include <pcms/coupler.h>
 #include <Omega_h_file.hpp>
 #include <Omega_h_for.hpp>
 #include <redev_variant_tools.h>
@@ -26,40 +25,42 @@ namespace ts = test_support;
 
 void xgc_delta_f(MPI_Comm comm, Omega_h::Mesh& mesh)
 {
-  pcms::Coupler cpl("proxy_couple_xgc_delta_f", comm);
+  pcms::Coupler coupler("proxy_couple", comm, false, {});
+  pcms::Application* app = coupler.AddApplication("proxy_couple_xgc_delta_f");
 
   auto is_overlap = ts::markOverlapMeshEntities(mesh, ts::IsModelEntInOverlap{});
-  cpl.AddField("gids",
+  app->AddField("gids",
                OmegaHFieldAdapter<GO>("global", mesh, is_overlap));
-  cpl.AddField("gids2",
+  app->AddField("gids2",
                OmegaHFieldAdapter<GO>("global", mesh, is_overlap));
   do {
     for (int i = 0; i < COMM_ROUNDS; ++i) {
-      cpl.BeginSendPhase();
-      cpl.SendField("gids");  //(Alt) df_gid_field->Send();
-      cpl.SendField("gids2"); //(Alt) df_gid_field->Send();
-      cpl.EndSendPhase();
-      cpl.BeginReceivePhase();
-      cpl.ReceiveField("gids"); //(Alt) df_gid_field->Receive();
-      cpl.EndReceivePhase();
+      app->BeginSendPhase();
+      app->SendField("gids");  //(Alt) df_gid_field->Send();
+      app->SendField("gids2"); //(Alt) df_gid_field->Send();
+      app->EndSendPhase();
+      app->BeginReceivePhase();
+      app->ReceiveField("gids"); //(Alt) df_gid_field->Receive();
+      app->EndReceivePhase();
       // cpl.ReceiveField("gids2"); //(Alt) df_gid_field->Receive();
     }
   } while (!done);
 }
 void xgc_total_f(MPI_Comm comm, Omega_h::Mesh& mesh)
 {
-  pcms::Coupler cpl("proxy_couple_xgc_total_f", comm);
+  pcms::Coupler coupler("proxy_couple", comm, false, {});
+  pcms::Application* app = coupler.AddApplication("proxy_couple_xgc_total_f");
   auto is_overlap = ts::markOverlapMeshEntities(mesh, ts::IsModelEntInOverlap{});
-  cpl.AddField("gids",
+  app->AddField("gids",
                OmegaHFieldAdapter<GO>("global", mesh, is_overlap));
   do {
     for (int i = 0; i < COMM_ROUNDS; ++i) {
-      cpl.BeginSendPhase();
-      cpl.SendField("gids"); //(Alt) tf_gid_field->Send();
-      cpl.EndSendPhase();
-      cpl.BeginReceivePhase();
-      cpl.ReceiveField("gids"); //(Alt) tf_gid_field->Receive();
-      cpl.EndReceivePhase();
+      app->BeginSendPhase();
+      app->SendField("gids"); //(Alt) tf_gid_field->Send();
+      app->EndSendPhase();
+      app->BeginReceivePhase();
+      app->ReceiveField("gids"); //(Alt) tf_gid_field->Receive();
+      app->EndReceivePhase();
     }
   } while (!done);
 }
@@ -68,7 +69,7 @@ void xgc_coupler(MPI_Comm comm, Omega_h::Mesh& mesh, std::string_view cpn_file)
   // coupling server using same mesh as application
   // note the xgc_coupler stores a reference to the internal mesh and it is the
   // user responsibility to keep it alive!
-  pcms::CouplerServer cpl(
+  pcms::Coupler cpl(
     "proxy_couple", comm, true,
     redev::Partition{ts::setupServerPartition(mesh, cpn_file)});
   const auto partition = std::get<redev::ClassPtn>(cpl.GetPartition());
